Log resource api times to statsd
Bug: 246953799

Log resource api times to statsd.  Two APIs are logged:
getResourceValue() and retrieveAttributes().  Other APIs may be timed
by the runtime layer but these are not logged.

This fixes a bug in ResourceTimer.startTimer() that caused the class
to skip the first repeated publication point.

This change renames the two RecourceTimer instances in
android_util_AssetManager.cpp to "_timer" to be more explicit and to
avoid the use of the more generic "_tag".

Test: manual
Change-Id: I0c5e0906f79ff35db1e34593275b39fb7f2f179b