Make IMMS#updateStatusIcon() multi-user aware
It turns out that

  InputMethodManagerService#updateStatusIcon()

still uses PakcageManager for user 0 when loading the IME application
label.  This CL fixes such a user mismatch.

Note that the code in question was originally added [1] before
multi-user support was added into InputMethodManagerService. We just
forgot to update the code when making IMMS multi-user aware [2].

 [1]: I1c1bd123d828fb10911bca92130e9a05c1f020b3
      b51b1656d370da98d68f72492db5442a64397152
 [2]: Ib23849d352db33f0747aa9d5a178f00ac726c13b
      b67143b01df1c4573ff5145cf4eec04c5098e85d

Fix: 250716279
Test: presubmit
Change-Id: I37106c3c82950d4f6f3a7b149b21579b0314a287