Resize tasks from corners when in touch mode
Update task resizing logic for desktop mode.
Add a new 44x44dp touch region in each corner of the task that is
centered around the corner. This is only used when resizing via touch.
This simplifies task resizing with touches and ensures that task resize
and back gesture do not conflict.

New task resize regions:
- with touch: use 44x44dp rect in each corner
- with mouse: use 30dp wide task edges

TODO:
- investigate how to pass taps in touch corners through to app

Bug: 251270585
Test: using touches to resize task for corners, check that resizing from
  edges no longer possible
Test: use mouse to resize task from edges

Change-Id: Ib1aaf69d2971c29a35200b0f3807871343aa1bf8