Bottom area buttons are LaunchableView
This is needed to make sure that the visibility of the views stays
consistent throughout the transition animation that ensues when the
views are clicked by the user to open an activity or dialog.

The next step is to actually add support for LaunchableView in
ActivityLaunchAnimator like we did for DialogLaunchAnimator.

Bug: 249986783
Test: clicked on the buttons, the animations still work, the linked bug
is still present (it's expected to be fixed when we add support in
ActivityLaunchAnimator)

Change-Id: I2e846545cb1f268fc007309cde7a4dea95bd408c