Stop using notification for IME switcher
InputMethodManagerService has had logic to show notifications when the
IME switcher icon cannot be shown on NavigationBar[1][2][3][4][5].

With this CL, such a fallback behavior is completely removed.  Devices
and form-factors that need a special handling need to come up with
their own solution rather than relying on this historical notification
mechanism.

 [1]: I2b9e955003c1f6703978824c947d5f322a039714
      01b6f164734fbc594c2c6d6c4f0747beb4f72866
 [2]: Id36c8c34159bea8b72557b40bcf024d401f580b6
      81bc8c1de06e55ec84a680275817e87975de5250
 [3]: I34a95732ef3e7c20d6549b57230c11f0c3db04d6
      2ce704453a30db7a1ab373a42b054f02cd075004
 [4]: I58421143bee25be459f4f744688b37614a750051
      4f8a92f9816d45553f8f8e5b7c9e11808f97d396
 [5]: I6ad818fc9ef08991e24c5020498243f1fc417998
      e155b92079eab3295cbb7c6b86cf6cf39bb00a2c

Bug: 245989733
Test: atest CtsPermission2TestCases:ProtectedBroadcastsTest
Test: presubmit
Change-Id: I3b09d2b05ba9bde686b360b9f720b5d4258551c3