Call IMM#startInputInner() after releasing the lock
While IMM#startInputInner() should not be called while the caller
already holds IMM#mH lock, it seems that there is a long standing
violation of such a rule for more than 10 years [1].

This CL finally addresses it.

The method calling order remains to be the same so hopefully the app
compat risk is low.

 [1]: I9189e6d2ac90aa8c621fdb44989728101e00329d
      382dacb66d89932b3fe6a37a43aa223f67a2003b

Bug: 234882948
Test: presubmit
Change-Id: Ib6f5a356ce5574976baefbe89e82de183b691e91