Fix a crash during mock testing
Bug: 249161784

This fixes a crash during mock testing of the broadcast queue when
BOOT_COMPLETED broadcasts are simulated.  The umInternal variable is
null, and umInternal.getUserInfo() throws a null pointer
exception. The change avoids dereferencing umInternal if it is null.

This was tested with a draft version of the broadcast replay code; it
was verified that BOOT_COMPLETED did not generate an NPE.

Test: atest
 * FrameworksMockingServicesTests:BroadcastQueueTest
Change-Id: I31f57f1dcb7f2ed9f02558f841e18958e8097373