Remove an unused param from IMMS#showInputMethodPickerFromClient()
This is the 2nd attempt to submit this cleanup [1], after a revert [2] due to some build failure.

This is a follow up CL to our previous CL [3], which introduced

  InputMethodManagerService#showInputMethodPickerFromClient(
          @NonNull IInputMethodClient client,
          int auxiliarySubtypeMode, int displayId);

Somehow the "client" has never been used, which is fine because the
actual caller verification is done by requiring WRITE_SECURE_SETTINGS
permission.

Let's remove such an unused parameter for simplicity.

There must be no observable behavior change.

 [1]: I2f27160cbe7d83201c8f52dbb6dcb774a5b8ae43
      4a435513cd7603880987fc71b5195041295abd84
 [2]: I343828ad83d50436211004753f1b9ba1b13b60ca
      c12e524f5ce38e98f6b2463d2dd2da0545d69a21
 [3]: Ic7d7c5a7ad8005a3fbd9d1c1b73e3c5a39a07001
      05de60e30f1c7644811868ed467c81cbbe6b0bf3

Bug: 234882948
Bug: 256952514
Test: presubmit
Test: lunch aosp_bramble-userdebug && make -j
Change-Id: Ia536d8793f0022179dfafd1cbf21091537375992