Implement setThreads APIs for PerformanceHintManager.Session.
Previously the list of threads of a Session was only determined when the
Session was created. This means newly forked threads from existing
threads of the Session will not get the benefit and the clients have to
create new Session for that.

This patch adds a new method to allow clients to update the threads of
the Session.

Bug:b/244216750
Test: atest PerformanceHintManagerTest
Test: atest PerformanceHintNativeTest
Test: atest HintManagerServiceTest

Change-Id: Iae8cbb4ce86a44a7cd9d6e68673c48800bed3a4e